Fix CORS issue.  Split server.py into additional db files.
The only endpoint requiring cross-site requests is request.

The db_options needs to be splitoff for eventual git-crypt usage.

I moved the db utility methods to another file just to clean things
up a little bit.

Updaged .gitignore to ignore the python cache.

Tested using localhost:5001 to send to receive via POST when
an additional server was running on localhost:5000.